app/src/main/java/com/hasanalic/ecommerce/feature_orders/domain/use_cases/GetOrderDetailUseCase.kt
package com.hasanalic.ecommerce.feature_orders.domain.use_cases

import com.hasanalic.ecommerce.core.domain.model.DataError
import com.hasanalic.ecommerce.core.domain.model.Result
import com.hasanalic.ecommerce.feature_orders.domain.model.OrderDetail
import com.hasanalic.ecommerce.feature_orders.domain.repository.OrderRepository
import javax.inject.Inject

class GetOrderDetailUseCase @Inject constructor(
    private val orderRepository: OrderRepository
) {
    suspend operator fun invoke(orderId: String): Result<OrderDetail, DataError.Local> {
        return orderRepository.getOrderDetail(orderId)
    }
}